+++ RELEASES.md
@@ -23,6 +23,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * service-templates-core: fixed issues, while generating screenshots for book
 * prov-template-compiler: PAST expression, detecting overloaded methods
 * prov-template-compiler: Python emitter, making use of overload annotation to disambiguate overloaded methods
+* prov-template-archetype: various improvements.
 
 
 `2.2.2`